Simplifies ffsea GO analyses

We use rrvgo, goSemSim and friends

go_simplify(
  x,
  method = NULL,
  ontology = c("BP", "MF", "CC"),
  direction = c("all", "up", "down"),
  similarity = c("Resnik", "Lin", "Rel", "Jiang", "Wang"),
  threshold = 0.7,
  max_padj = 0.1,
  min_logFC = NULL,
  score_by = NULL,
  semdata = NULL,
  ...,
  verbose = FALSE
)

Arguments

x
A FacileFseaAnalysisResult

ontology
which ontology are were summarizing? BP, MF, or CC

direction
focus on all genes, or only up or down regulated genes (works for ttest, not ANOVA)

similarity
The method to use to measure similarity, see here.

min_logFC, max_padj
parameters used to define signficant genes. If the TREAT framework was used, defaults to the threshold tested, otherwise its zero.