Aggregates gene-level stats from genes in leading edge.

The logFC and t-statistics of the genes from the leading edge (defined in extract_leading_edge()) are used to give an effect size of the geneset shift.

leading_edge_scores(
  x,
  leading_edge = NULL,
  aggregate = c("t", "logFC"),
  stats = NULL,
  ...
)

Arguments

x
The FacileFseaAnalysisResult objects where fgsea was ran.

leading_edge
Precaculated leading edge genes, as in the output from leading_edge(). If NULL, the leading_edge will be caluclated for all genesets tested x then scored. The scoring can take a long time if there are many, so you may want to provide a filtered leading_edge set.

aggregate
the column names from the dge results to summarize, defaults to c("t", "logFC")

stats
A name of a gsea method run in "x" to append pvalues from. If NULL (default), none will be added.

Value

a tibble of geneset scores, aggregated by leading edge genesets for columns in aggregate. If "fgsea" is a method run in "x", ES, and NES columns will be returned as well.

Details

This is used in place of the ES/NES mojo.